scene blending queue (issue #10), shader parameters applies only if it was changed, render loop performance remarks, sponza sample statistic update by timer

+++ lite3d/src/lite3d_framebuffer.c
@@ -26,7 +26,7 @@
 #include <lite3d/lite3d_framebuffer.h>
 
 static lite3d_framebuffer *gCurrentFb = NULL;
-static int gMaxColorAttachments = 0;
+static size_t gMaxColorAttachments = 0;
 static int gMaxFramebufferSize = 0;
 
 /*
